pyproject.toml
[project]
name = "pangram-breaker"
version = "0.1.0"
description = "Security research: testing AI text detection via Pangram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pangram-sdk>=0.1.11",
    "pydantic>=2.12.5",
    "tomli>=2.4.1",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pangram_breaker"]
